Convert to Multiplatform with node test and bintray publishing
