+++ gradle.properties
@@ -1,21 +1,31 @@
-# Project-wide Gradle settings.
-# IDE (e.g. Android Studio) users:
-# Gradle settings configured through the IDE *will override*
-# any settings specified in this file.
-# For more details on how to configure your build environment visit
-# http://www.gradle.org/docs/current/userguide/build_environment.html
-# Specifies the JVM arguments used for the daemon process.
-# The setting is particularly useful for tweaking memory settings.
-org.gradle.jvmargs=-Xmx1536m
-# When configured, Gradle will run in incubating parallel mode.
-# This option should only be used with decoupled projects. More details, visit
-# http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
-# org.gradle.parallel=true
-# AndroidX package structure to make it clearer which packages are bundled with the
-# Android operating system, and which are packaged with your app's APK
-# https://developer.android.com/topic/libraries/support-library/androidx-rn
-android.useAndroidX=true
-# Automatically convert third-party libraries to use AndroidX
-android.enableJetifier=true
-# Kotlin code style for this project: "official" or "obsolete":
+# Enable official Kotlin Code Style in the IDE.
 kotlin.code.style=official
+# Set up your maven coordinates here, artifactId is defined per project.
+version=0.0.1
+group=org.bisdk
+
+# Put link to your OSS repository. Bintray requires this to be filled for free OpenSource tiers.
+vcs=https://bisdk.org
+description=BiSDK
+# Put your Bintray organization name here, e.g. your bintray username.
+bintrayOrg=bisdk
+# Put your Bintray repository name here, e.g. maven.
+bintrayRepository=maven
+# Put name of your package here.
+bintrayPackage=org.bisdk.sdk
+# Dependencies versions
+kotlin_version=1.3.11
+junit_version=4.12
+gradle_node_version=1.2.0
+node_version=8.9.3
+npm_version=5.7.1
+mocha_version=6.0.2
+mocha_teamcity_reporter_version=2.5.2
+source_map_support_version=0.5.10
+# This is a patched version of bintray plugin, don't change it unless you know what you're doing.
+bintray_plugin_version=1.8.4
+kotlinx_io_version=0.1.6
+# If you want to test publishing, add these two lines to `local.properties` file next to this one.
+# That file is not git-controlled, so you can put your bintray credentials there.
+# Obviously, remove # comment at the beginning of the line.
+
