+++ gradle/pom.gradle
@@ -1,0 +1,37 @@
+
+def pomConfig = {
+    licenses {
+        license([:]) {
+            name "MIT License"
+            url "http://opensource.org/licenses/MIT"
+            distribution "repo"
+        }
+    }
+    developers {
+        developer {
+            id "bisdk"
+            name "Stefan Wärting"
+            organization "bisdk"
+            organizationUrl "https://bisdk.org"
+        }
+    }
+
+    scm {
+        url 'https:///bisdk.org'
+    }
+}
+
+project.ext.configureMavenCentralMetadata = { pom ->
+
+    println("configureMavenCentralMetadata")
+    println("project.name: $project.name")
+    println("project.description: $project.description")
+    println("project.vcs: $project.vcs")
+    def root = asNode()
+    root.appendNode('name', project.name)
+    root.appendNode('description', project.description)
+    root.appendNode('url', project.vcs)
+    root.children().last() + pomConfig
+}
+
+project.ext.configurePom = pomConfig
