+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,4 +1,3 @@
-#Thu Feb 07 19:31:07 CET 2019
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStoreBase=GRADLE_USER_HOME
