+++ local.properties.example
@@ -1,0 +1,2 @@
+# bintrayUser=username
+# bintrayApiKey=apikey
