--- projects-sdk.gradle
@@ -1,1 +1,0 @@
-include ':sdk'
