--- sdk/build.gradle
@@ -1,19 +1,0 @@
-apply plugin: 'java-library'
-apply plugin: 'kotlin'
-
-
-dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation 'org.jetbrains.kotlin:kotlin-reflect:1.3.21'
-    implementation 'com.fasterxml.jackson.core:jackson-databind:2.9.8'
-    implementation 'com.fasterxml.jackson.module:jackson-module-kotlin:2.9.8'
-    implementation 'com.fasterxml.jackson.module:jackson-module-parameter-names:2.9.8'
-    testImplementation 'org.jetbrains.kotlin:kotlin-test-junit:1.3.11'
-    testImplementation 'junit:junit:4.12'
-    testImplementation 'org.assertj:assertj-core:3.9.1'
-}
-
-sourceCompatibility = "7"
-targetCompatibility = "7"
