--- sdk/src/main/java/org/bisdk/sdk/AuthenticationException.kt
@@ -1,3 +1,0 @@
-package org.bisdk.sdk
-
-class AuthenticationException(s: String) : Exception(s)
