+++ src/commonMain/kotlin/org/bisdk/AuthenticationException.kt
@@ -1,0 +1,3 @@
+package org.bisdk
+
+class AuthenticationException(s: String) : Exception(s)
