+++ src/commonMain/kotlin/org/bisdk/BiSdk.kt
@@ -1,0 +1,5 @@
+package org.bisdk
+
+class BiSdk {
+
+}
