+++ src/commonMain/kotlin/org/bisdk/Sample.kt
@@ -1,0 +1,11 @@
+package org.bisdk
+
+expect class Sample() {
+    fun checkMe(): Int
+}
+
+expect object Platform {
+    val name: String
+}
+
+fun hello(): String = "Hello from ${Platform.name}"
