+++ src/commonTest/kotlin/org/bisdk/SampleTests.kt
@@ -1,0 +1,11 @@
+package org.bisdk
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class SampleTests {
+    @Test
+    fun testMe() {
+        assertTrue(Sample().checkMe() > 0)
+    }
+}
