+++ src/jsMain/kotlin/org/bisdk/SampleJs.kt
@@ -1,0 +1,9 @@
+package org.bisdk
+
+actual class Sample {
+    actual fun checkMe() = 12
+}
+
+actual object Platform {
+    actual val name: String = "JS"
+}
