+++ src/jsTest/kotlin/org/bisdk/SampleTestsJS.kt
@@ -1,0 +1,11 @@
+package org.bisdk
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class SampleTestsJS {
+    @Test
+    fun testHello() {
+        assertTrue("JS" in hello())
+    }
+}
