+++ src/jvmMain/kotlin/org/bisdk/sdk/Encoding.kt
