+++ src/jvmMain/kotlin/org/bisdk/sdk/Transition.kt
