+++ src/jvmTest/kotlin/org/bisdk/SampleTestsJVM.kt
@@ -1,0 +1,11 @@
+package org.bisdk
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class SampleTestsJVM {
+    @Test
+    fun testHello() {
+        assertTrue("JVM" in hello())
+    }
+}
