+++ src/jvmTest/kotlin/org/bisdk/sdk/EncodingKtTest.kt
