+++ src/jvmTest/kotlin/org/bisdk/sdk/StartupTest.kt
