+++ src/jvmTest/kotlin/org/bisdk/sdk/TransitionTest.kt
