+++ src/macosMain/kotlin/org/bisdk/SampleMacos.kt
@@ -1,0 +1,9 @@
+package org.bisdk
+
+actual class Sample {
+    actual fun checkMe() = 7
+}
+
+actual object Platform {
+    actual val name: String = "Native"
+}
