+++ src/macosTest/kotlin/org/bisdk/SampleTestsNative.kt
@@ -1,0 +1,11 @@
+package org.bisdk
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class SampleTestsNative {
+    @Test
+    fun testHello() {
+        assertTrue("Native" in hello())
+    }
+}
